[fix][req]接口表数据复制bugfix

See merge request scm-pms/scm-wip-backstage!53

+++ wip-infrastructure/src/main/java/com/cvte/scm/wip/infrastructure/requirement/mapper/dataobject/XxwipMoInterfaceDO.java
@@ -55,7 +55,7 @@ public class XxwipMoInterfaceDO {
 
     @Column(name = "ITEM_QTY")
     @ApiModelProperty(value = "物料数量")
-    private Integer itemQty;
+    private Long itemQty;
 
     @Column(name = "WKP_NO")
     @ApiModelProperty(value = "工序号")
